.classpath: auto generated? .gitignore: ignore .storage, kill.sh, env/*/config/hosts.config,     env/*/config/system.config, event.txt, rungui batchp2p.py: run 200 nodes in mininet BlockChainDemo2.iml: remove file://$MODULE_DIR$/src/main/resources enter.sh: enter network namespace by pid Makefile: add blockinit, runguibuild, batchkill rungui.cpp: rungui with correct network namespace and env directory settings.json: vscode watcherExclude **/env/** Main.java: run gui with 2 command args ServerMain.java: if no block_data file, clean db and java vars,     otherwise loadvar. Sync blockchain status after server started ContractServer.java: set Tool.currentClientID=msgCtx.getSender();     add timestamp in network message,     add process getState, store var status, add getSnapshot__,     add installSnapshot__ Equip_Contract.java: unify getCurrentTime() Initialization_Contract.java: unify getLocalIp(), getCurrentTime() Res_Contract.java: unify getCurrentTime() Tool.java: add maxClientCount=200, add timestamp in network message,     unify getLocalIp(), add system(), add levelDBClear(),     add writeXXXXDB with ArrayList arg, add storeVar(), add loadVar(),     unify getClientIP(), unify getCurrentTime() User_Contract.java: unify getClientIP(), unify getCurrentTime(),     solve getUsereds display "-1" problem ApplyMessage.java: implements Serializable Equip.java: implements Serializable MiscCommand.java: add getState command Resource.java: implements Serializable User.java: implements Serializable ApplyRoleUI.java: not allowed to change role to "管理节点" LevelDBUtilLocal.java: dbFolder with no nodeID

+++ src/main/java/Communication/ServerMain.java
@@ -1,20 +1,62 @@
 package Communication;
+import java.io.ByteArrayOutputStream;
 import java.io.FileOutputStream;
+import java.io.ObjectOutput;
+import java.io.ObjectOutputStream;
 import java.io.PrintStream;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.security.MessageDigest;
+import java.util.Base64;
 
+import javax.swing.plaf.basic.BasicInternalFrameTitlePane.SystemMenuBar;
+
+import bftsmart.tom.ServiceProxy;
 import contract.Contracts.ContractServer;
+import contract.Contracts.Initialization_Contract;
 import contract.Contracts.Tool;
+import contract.Struct.MiscCommand;
 
 public class ServerMain {
     public static void main(String[] args) {
         Tool.nodeID=Integer.parseInt(args[0]);
-        String destination = "block_data";
-        try(PrintStream ps = new PrintStream(new FileOutputStream(destination, false))){
-            ps.printf("Block data\n\n\n");
+        String destination = Tool.blockPath; 
+        // if there is no,block data file, init the block chain
+        // otherwise, no init operations
+        if(!Files.exists(Paths.get(destination))){
+            try(PrintStream ps = new PrintStream(new FileOutputStream(destination, false))){
+                ps.printf("Block data\n\n\n");
+            }
+            catch(Exception e){
+                e.printStackTrace();
+            }
+            System.out.println("#################clear db!!!!!!!!!!!!!");
+            Tool.levelDBClear();
+            Tool.storeVar();
+        }
+        else{
+            Tool.loadVar();
+        }
+        ContractServer contractServer=new ContractServer(Tool.nodeID);
+        Tool.initSnapshot=contractServer.getSnapshot__();
+        ServiceProxy c=new ServiceProxy(Tool.nodeID);
+        try (ByteArrayOutputStream byteOut = new ByteArrayOutputStream();
+        ObjectOutput objOut = new ObjectOutputStream(byteOut);) {
+            System.out.println("Wait servers to be ready.....");
+            Thread.sleep(10*1000);
+            System.out.println("Wait done......");
+            objOut.writeObject(MiscCommand.getState);
+            objOut.flush();
+            byteOut.flush();
+            byte[] reply = c.invokeOrdered(byteOut.toByteArray());
+            MessageDigest md = MessageDigest.getInstance("SHA-256");
+            md.update(reply);
+            System.out.printf("Returned hash: %s\n",Base64.getEncoder().encodeToString(md.digest()));
+            contractServer.installSnapshot__(reply);
+            c.close();
         }
         catch(Exception e){
             e.printStackTrace();
         }
-        ContractServer contractServer=new ContractServer(Tool.nodeID);
     }
 }
